requirements.txt
easydict
ffmpeg-python
gradio
omegaconf
onnxruntime==1.15.1
mmdeploy==1.2.0
mmdeploy-runtime==1.2.0
scikit-image
tqdm